--- kwon-dae-won/app/src/main/java/kwon/dae/won/FlickrScreen.kt
@@ -1,83 +1,0 @@
-package kwon.dae.won
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
-import androidx.lifecycle.viewmodel.compose.viewModel
-import androidx.paging.compose.collectAsLazyPagingItems
-import androidx.paging.compose.items
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.unit.dp
-import androidx.paging.compose.LazyPagingItems
-import coil.compose.AsyncImagePainter
-import coil.compose.rememberAsyncImagePainter
-import kwon.dae.won.domain.model.Photo
-
-
-/**
- * @author Daewon on 10,February,2023
- *
- */
-@Composable
-fun FlickrScreen(photos: LazyPagingItems<Photo>) {
-
-
-    LazyColumn {
-        items(
-            items = photos
-        ) { photo ->
-            photo?.let {
-                Row(
-                    horizontalArrangement = Arrangement.Center,
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    val serverId = 65535
-                    val url = "https://live.staticflickr.com/${serverId}"
-                    if (photo.id.isNotBlank()) {
-                        var isImageLoading by remember { mutableStateOf(false) }
-
-                        val painter = rememberAsyncImagePainter(
-                            model = url + "/${photo.id}_${photo.secret}.jpg"
-                        )
-
-                        isImageLoading = when (painter.state) {
-                            is AsyncImagePainter.State.Loading -> true
-                            else -> false
-                        }
-
-                        Box(
-                            contentAlignment = Alignment.Center
-                        ) {
-                            Image(
-                                modifier = Modifier
-                                    .height(150.dp)
-                                    .width(300.dp),
-                                painter = painter,
-                                contentDescription = "Photo Image",
-                                contentScale = ContentScale.FillBounds,
-                            )
-
-                            if (isImageLoading) {
-                                CircularProgressIndicator(
-                                    modifier = Modifier
-                                        .padding(horizontal = 6.dp, vertical = 3.dp),
-                                    color = MaterialTheme.colorScheme.primary
-                                )
-                            }
-                        }
-
-                    }
-                }
-            }
-        }
-    }
-}
